Correcciones en la tabla de producto el campo initialWeight que solo aceptaba 5 digitos totales con 2 decimales, modificado por 6 totales. Correccion de endpoint Productos para que primero verifique QueryParams, luego Params y sin parametros.

+++ src/products/model/entities/product.entity.ts
@@ -23,7 +23,7 @@ export class Product extends Model {
     @Column
     type: string;
 
-    @Column({ type: DataType.FLOAT(4, 2) })
+    @Column({ type: DataType.FLOAT(6, 2) })
     initialWeight: number;
